fix(schema): cash_account_transactions FK type + Upbit env docs
문제 1 — schema.sql 적용 시:
  [HY000][3780] Referencing column 'cash_account_id' and referenced column 'id'
  in foreign key constraint 'fk_cash_tx_cash_account_id' are incompatible.

원인: cash_account_transactions.cash_account_id 가 BIGINT 인데
cash_accounts.id 는 INT — MySQL FK 타입 불일치로 거부.

수정:
- alembic 마이그레이션 d5e8d86ff010: id/cash_account_id 를 BigInteger → Integer.
  (이 revision 은 production 에 미적용 상태라 직접 수정 안전.
   다른 모든 테이블의 id 컬럼도 INT 라 일관성 회복.)
- schema.sql: BIGINT → INT 동일.
- 모델 Mapped[int] 는 기본 Integer 라 변경 불필요.

문제 2 — Upbit 동기화 환경변수 README 미안내:
  PR #125 가 sync UI 만 추가, 키 설정 안내가 root README/.env.example 누락.

수정:
- backend/.env.example: UPBIT_ACCESS_KEY/SECRET_KEY 섹션 추가 + 권한 안내.
- README.md: 로컬 setup + production env 섹션에 Upbit 키 안내 추가.

+++ backend/alembic/versions/d5e8d86ff010_create_cash_account_transactions.py
@@ -20,8 +20,8 @@ depends_on: str | Sequence[str] | None = None
 def upgrade() -> None:
     op.create_table(
         "cash_account_transactions",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
-        sa.Column("cash_account_id", sa.BigInteger(), nullable=True),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+        sa.Column("cash_account_id", sa.Integer(), nullable=True),
         sa.Column(
             "kind",
             sa.Enum(
